rename notebook.ts to cell.ts
